add AbstractEventStoreReadModelProjectionTest
