+++ tests/Projection/InMemoryEventStoreReadModelProjectionTest.php
@@ -12,33 +12,25 @@ declare(strict_types=1);
 
 namespace ProophTest\EventStore\Projection;
 
-use ArrayIterator;
-use PHPUnit\Framework\TestCase;
-use Prooph\Common\Messaging\Message;
 use Prooph\EventStore\EventStore;
 use Prooph\EventStore\EventStoreDecorator;
 use Prooph\EventStore\Exception\InvalidArgumentException;
-use Prooph\EventStore\Exception\RuntimeException;
 use Prooph\EventStore\InMemoryEventStore;
 use Prooph\EventStore\Projection\InMemoryEventStoreReadModelProjection;
 use Prooph\EventStore\Projection\InMemoryProjectionManager;
-use Prooph\EventStore\Stream;
-use Prooph\EventStore\StreamName;
 use ProophTest\EventStore\Mock\ReadModelMock;
-use ProophTest\EventStore\Mock\UserCreated;
-use ProophTest\EventStore\Mock\UsernameChanged;
 
-class InMemoryEventStoreReadModelProjectionTest extends TestCase
+class InMemoryEventStoreReadModelProjectionTest extends AbstractEventStoreReadModelProjectionTest
 {
     /**
      * @var InMemoryProjectionManager
      */
-    private $projectionManager;
+    protected $projectionManager;
 
     /**
      * @var InMemoryEventStore
      */
-    private $eventStore;
+    protected $eventStore;
 
     protected function setUp(): void
     {
@@ -49,581 +41,57 @@ class InMemoryEventStoreReadModelProjectionTest extends TestCase
     /**
      * @test
      */
-    public function it_can_project_from_stream_and_reset(): void
+    public function it_throws_exception_when_trying_to_run_two_projections_at_the_same_time(): void
     {
-        $this->prepareEventStream('user-123');
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $this->assertEquals('test_projection', $projection->getName());
-
-        $projection
-            ->init(function (): array {
-                return ['count' => 0];
-            })
-            ->fromStream('user-123')
-            ->when([
-                UsernameChanged::class => function (array $state, UsernameChanged $event): array {
-                    $state['count']++;
-
-                    return $state;
-                },
-            ])
-            ->run(false);
-
-        $this->assertEquals(49, $projection->getState()['count']);
-
-        $projection->reset();
-
-        $projection->run(false);
-
-        $projection->run(false);
-
-        $this->assertEquals(49, $projection->getState()['count']);
+        $this->markTestSkipped('InMemoryProjectionManager cannot guard agains concurrent projections');
     }
 
     /**
      * @test
      */
-    public function it_can_project_from_stream_and_delete(): void
+    public function it_deletes_when_projection_before_start_when_it_was_deleted_from_outside(): void
     {
-        $this->prepareEventStream('user-123');
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $this->assertEquals('test_projection', $projection->getName());
-
-        $projection
-            ->init(function (): array {
-                return ['count' => 0];
-            })
-            ->fromStream('user-123')
-            ->when([
-                UsernameChanged::class => function (array $state, UsernameChanged $event): array {
-                    $state['count']++;
-
-                    return $state;
-                },
-            ])
-            ->run(false);
-
-        $this->assertEquals(49, $projection->getState()['count']);
-
-        $projection->delete(true);
-
-        $this->assertFalse($readModel->isInitialized());
+        $this->markTestSkipped('InMemoryProjectionManager cannot delete projections');
     }
 
     /**
      * @test
      */
-    public function it_can_be_stopped_while_processing(): void
+    public function it_deletes_projection_during_run_when_it_was_deleted_from_outside(): void
     {
-        $this->prepareEventStream('user-123');
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection
-            ->init(function (): array {
-                return ['count' => 0];
-            })
-            ->fromStream('user-123')
-            ->whenAny(function (array $state, Message $event): array {
-                $state['count']++;
-
-                if ($state['count'] === 10) {
-                    $this->stop();
-                }
-
-                return $state;
-            })
-            ->run(false);
-
-        $this->assertEquals(10, $projection->getState()['count']);
+        $this->markTestSkipped('InMemoryProjectionManager cannot delete projections');
     }
 
     /**
      * @test
      */
-    public function it_can_query_from_streams(): void
+    public function it_resets_projection_before_start_when_it_was_reset_from_outside(): void
     {
-        $this->prepareEventStream('user-123');
-        $this->prepareEventStream('user-234');
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection
-            ->init(function (): array {
-                return ['count' => 0];
-            })
-            ->fromStreams('user-123', 'user-234')
-            ->whenAny(
-                function (array $state, Message $event): array {
-                    $state['count']++;
-
-                    return $state;
-                }
-            )
-            ->run(false);
-
-        $this->assertEquals(100, $projection->getState()['count']);
+        $this->markTestSkipped('InMemoryProjectionManager cannot reset projections');
     }
 
     /**
      * @test
      */
-    public function it_can_query_from_all_ignoring_internal_streams(): void
+    public function it_resets_projection_during_run_when_it_was_reset_from_outside(): void
     {
-        $this->prepareEventStream('user-123');
-        $this->prepareEventStream('user-234');
-        $this->prepareEventStream('$iternal-345');
-
-        $testCase = $this;
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection
-            ->init(function (): array {
-                return ['count' => 0];
-            })
-            ->fromAll()
-            ->whenAny(
-                function (array $state, Message $event) use ($testCase): array {
-                    $state['count']++;
-                    if ($state['count'] < 51) {
-                        $testCase->assertEquals('user-123', $this->streamName());
-                    } else {
-                        $testCase->assertEquals('user-234', $this->streamName());
-                    }
-
-                    return $state;
-                }
-            )
-            ->run(false);
-
-        $this->assertEquals(100, $projection->getState()['count']);
+        $this->markTestSkipped('InMemoryProjectionManager cannot reset projections');
     }
 
     /**
      * @test
      */
-    public function it_can_query_from_category_with_when_all(): void
+    public function it_stops_when_projection_before_start_when_it_was_stopped_from_outside(): void
     {
-        $this->prepareEventStream('user-123');
-        $this->prepareEventStream('user-234');
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection
-            ->init(function (): array {
-                return ['count' => 0];
-            })
-            ->fromCategory('user')
-            ->whenAny(
-                function (array $state, Message $event): array {
-                    $state['count']++;
-
-                    return $state;
-                }
-            )
-            ->run(false);
-
-        $this->assertEquals(100, $projection->getState()['count']);
+        $this->markTestSkipped('InMemoryProjectionManager cannot stop projections');
     }
 
     /**
      * @test
      */
-    public function it_can_query_from_categories_with_when(): void
+    public function it_stops_projection_during_run_when_it_was_stopped_from_outside(): void
     {
-        $this->prepareEventStream('user-123');
-        $this->prepareEventStream('user-234');
-        $this->prepareEventStream('guest-345');
-        $this->prepareEventStream('guest-456');
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection
-            ->init(function (): array {
-                return ['count' => 0];
-            })
-            ->fromCategories('user', 'guest')
-            ->when([
-                UserCreated::class => function (array $state, Message $event): array {
-                    $state['count']++;
-
-                    return $state;
-                },
-            ])
-            ->run(false);
-
-        $this->assertEquals(4, $projection->getState()['count']);
-    }
-
-    public function it_resumes_projection_from_position(): void
-    {
-        $this->prepareEventStream('user-123');
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection
-            ->init(function (): array {
-                return ['count' => 0];
-            })
-            ->fromCategories('user', 'guest')
-            ->when([
-                UsernameChanged::class => function (array $state, Message $event): array {
-                    $state['count']++;
-
-                    return $state;
-                },
-            ])
-            ->run(false);
-
-        $this->assertEquals(49, $projection->getState()['count']);
-
-        $events = [];
-        for ($i = 51; $i <= 100; $i++) {
-            $events[] = UsernameChanged::with([
-                'name' => uniqid('name_'),
-            ], $i);
-        }
-
-        $this->eventStore->appendTo(new StreamName('user-123'), new ArrayIterator($events));
-
-        $projection->run();
-
-        $this->assertEquals(99, $projection->getState()['count']);
-    }
-
-    /**
-     * @test
-     */
-    public function it_resets_to_empty_array(): void
-    {
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $state = $projection->getState();
-
-        $this->assertInternalType('array', $state);
-
-        $projection->reset();
-
-        $state2 = $projection->getState();
-
-        $this->assertInternalType('array', $state2);
-    }
-
-    /**
-     * @test
-     */
-    public function it_throws_exception_when_init_callback_provided_twice(): void
-    {
-        $this->expectException(RuntimeException::class);
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection->init(function (): array {
-            return [];
-        });
-        $projection->init(function (): array {
-            return [];
-        });
-    }
-
-    /**
-     * @test
-     */
-    public function it_throws_exception_when_from_called_twice(): void
-    {
-        $this->expectException(RuntimeException::class);
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection->fromStream('foo');
-        $projection->fromStream('bar');
-    }
-
-    /**
-     * @test
-     */
-    public function it_throws_exception_when_from_called_twice_2(): void
-    {
-        $this->expectException(RuntimeException::class);
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection->fromStreams('foo');
-        $projection->fromCategory('bar');
-    }
-
-    /**
-     * @test
-     */
-    public function it_throws_exception_when_from_called_twice_3(): void
-    {
-        $this->expectException(RuntimeException::class);
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection->fromCategory('foo');
-        $projection->fromStreams('bar');
-    }
-
-    /**
-     * @test
-     */
-    public function it_throws_exception_when_from_called_twice_4(): void
-    {
-        $this->expectException(RuntimeException::class);
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection->fromCategories('foo');
-        $projection->fromCategories('bar');
-    }
-
-    /**
-     * @test
-     */
-    public function it_throws_exception_when_from_called_twice_5(): void
-    {
-        $this->expectException(RuntimeException::class);
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection->fromCategories('foo');
-        $projection->fromAll('bar');
-    }
-
-    /**
-     * @test
-     */
-    public function it_throws_exception_when_when_called_twice_(): void
-    {
-        $this->expectException(RuntimeException::class);
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection->when(['foo' => function (): void {
-        }]);
-        $projection->when(['foo' => function (): void {
-        }]);
-    }
-
-    /**
-     * @test
-     */
-    public function it_throws_exception_when_invalid_handlers_configured(): void
-    {
-        $this->expectException(InvalidArgumentException::class);
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection->when(['1' => function (): void {
-        }]);
-    }
-
-    /**
-     * @test
-     */
-    public function it_throws_exception_when_invalid_handlers_configured_2(): void
-    {
-        $this->expectException(InvalidArgumentException::class);
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection->when(['foo' => 'invalid']);
-    }
-
-    /**
-     * @test
-     */
-    public function it_throws_exception_when_whenAny_called_twice(): void
-    {
-        $this->expectException(RuntimeException::class);
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection->whenAny(function (): void {
-        });
-        $projection->whenAny(function (): void {
-        });
-    }
-
-    /**
-     * @test
-     */
-    public function it_throws_exception_on_run_when_nothing_configured(): void
-    {
-        $this->expectException(RuntimeException::class);
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-        $projection->run();
-    }
-
-    /**
-     * @test
-     */
-    public function it_updates_read_model_using_when(): void
-    {
-        $this->prepareEventStream('user-123');
-
-        $testCase = $this;
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection
-            ->fromAll()
-            ->when([
-                UserCreated::class => function ($state, Message $event) use ($testCase): void {
-                    $testCase->assertEquals('user-123', $this->streamName());
-                    $this->readModel()->stack('insert', 'name', $event->payload()['name']);
-                },
-                UsernameChanged::class => function ($state, Message $event) use ($testCase): void {
-                    $testCase->assertEquals('user-123', $this->streamName());
-                    $this->readModel()->stack('update', 'name', $event->payload()['name']);
-
-                    if ($event->payload()['name'] === 'Sascha') {
-                        $this->stop();
-                    }
-                },
-            ])
-            ->run();
-
-        $this->assertEquals('Sascha', $readModel->read('name'));
-
-        $projection->reset();
-
-        $this->assertFalse($readModel->hasKey('name'));
-    }
-
-    /**
-     * @test
-     */
-    public function it_updates_read_model_using_when_any(): void
-    {
-        $this->prepareEventStream('user-123');
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel);
-
-        $projection
-            ->init(function (): void {
-                $this->readModel()->stack('insert', 'name', null);
-            })
-            ->fromStream('user-123')
-            ->whenAny(function ($state, Message $event): void {
-                $this->readModel()->stack('update', 'name', $event->payload()['name']);
-
-                if ($event->payload()['name'] === 'Sascha') {
-                    $this->stop();
-                }
-            })
-            ->run();
-
-        $this->assertEquals('Sascha', $readModel->read('name'));
-    }
-
-    /**
-     * @test
-     */
-    public function it_persists_in_single_handler(): void
-    {
-        $this->prepareEventStream('user-123');
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel, [
-            $this->projectionManager::OPTION_PERSIST_BLOCK_SIZE => 10,
-        ]);
-
-        $projection
-            ->init(function (): array {
-                return ['count' => 0];
-            })
-            ->fromCategories('user', 'guest')
-            ->whenAny(function (array $state, Message $event): array {
-                $state['count']++;
-
-                return $state;
-            })
-            ->run(false);
-
-        $this->assertEquals(50, $projection->getState()['count']);
-    }
-
-    /**
-     * @test
-     */
-    public function it_persists_in_handlers(): void
-    {
-        $this->prepareEventStream('user-123');
-
-        $readModel = new ReadModelMock();
-
-        $projection = $this->projectionManager->createReadModelProjection('test_projection', $readModel, [
-            $this->projectionManager::OPTION_PERSIST_BLOCK_SIZE => 10,
-        ]);
-
-        $projection
-            ->init(function (): array {
-                return ['count' => 0];
-            })
-            ->fromCategories('user', 'guest')
-            ->when([
-                UsernameChanged::class => function (array $state, Message $event): array {
-                    $state['count']++;
-
-                    return $state;
-                },
-            ])
-            ->run(false);
-
-        $this->assertEquals(49, $projection->getState()['count']);
+        $this->markTestSkipped('InMemoryProjectionManager cannot stop projections');
     }
 
     /**
@@ -680,23 +148,5 @@ class InMemoryEventStoreReadModelProjectionTest extends TestCase
         $wrappedEventStore->getInnerEventStore()->willReturn($eventStore->reveal())->shouldBeCalled();
 
         new InMemoryEventStoreReadModelProjection($wrappedEventStore->reveal(), 'test_projection', new ReadModelMock(), 1, 1, 1);
-    }
-
-    private function prepareEventStream(string $name): void
-    {
-        $events = [];
-        $events[] = UserCreated::with([
-            'name' => 'Alex',
-        ], 1);
-        for ($i = 2; $i < 50; $i++) {
-            $events[] = UsernameChanged::with([
-                'name' => uniqid('name_'),
-            ], $i);
-        }
-        $events[] = UsernameChanged::with([
-            'name' => 'Sascha',
-        ], 50);
-
-        $this->eventStore->create(new Stream(new StreamName($name), new ArrayIterator($events)));
     }
 }
